pyproject.toml
[tool.poetry]
name = "rti-tracker"
version = "1"
description = "Tracks the status of RTIs filed by SFLC.in"
authors = ["Nikhil Kumar Singh <nk_mason@protonmail.com>"]
license = "GPL-3.0"

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.75.0"
uvicorn = {extras = ["standard"], version = "^0.17.6"}
toml = "^0.10.2"
tortoise-orm = {extras = ["asyncpg"], version = "^0.18.1"}
ciso8601 = "^2.2.0"
aerich = "^0.6.2"
python-multipart = "^0.0.5"
aiosqlite = "^0.17.0"
email-validator = "^1.1.3"
lxml = "^4.8.0"
requests = "^2.27.1"
python-slugify = "^6.1.1"
passlib = "^1.7.4"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
Pillow = "^9.0.1"
pytesseract = "^0.3.9"
watchgod = "^0.8.1"
aiofiles = "^0.8.0"
typer = {extras = ["all"], version = "^0.4.1"}
fastapi-pagination = {extras = ["tortoise-orm"], version = "^0.9.3"}

[tool.poetry.dev-dependencies]
black = {version = "^22.1.0", allow-prereleases = true}
Faker = "^13.3.5"
tqdm = "^4.64.0"
fastapi-profiler = "^1.0.0"


[tool.aerich]
tortoise_orm = "aerich_config.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
